fix(pos-online): address Phase 3 review M1, M2, M6
M1: Add .passthrough() to 6 response wrapper schemas (DraftOrder/PriceRule/
    DiscountCode list+single) — was inconsistent with other 9 resources
M2: customer.total_spent z.number().int() -> z.number() (Sapo uses FLOAT
    VND across ALL money fields, including total_spent)
M6: get_variant description "integer VND" -> "VND, float with 4 decimals"

Defer M3/M4/M5 (handcrafted customer/variant/order_return fixtures) until
real data available on dev store. Defer M7 (delete_strict classification
nuance for delete_price_rule). Tracked in Phase 5 release readiness.

+++ src/schemas/customer.ts
@@ -7,7 +7,7 @@
  * - gender field: Male | Female | Other (Sapo-specific)
  * - dob: birth date YYYY-MM-DD (Sapo-specific)
  * - state enum: enabled | disabled | invited | decline
- * - total_spent: integer VND (not decimal)
+ * - total_spent: FLOAT VND (Sapo serializes money with 4 decimals across all resources)
  * - addresses nested; see AddressSchema
  */
 
@@ -38,7 +38,7 @@ export const CustomerSchema = z
     dob: z.string().optional().nullable(),
     orders_count: z.number().int().optional(),
     /** Total amount spent in integer VND */
-    total_spent: z.number().int().optional(),
+    total_spent: z.number().optional(),
     tags: z.string().optional().nullable(),
     note: z.string().optional().nullable(),
     last_order_id: z.number().int().optional().nullable(),
